refactor to student review

+++ src/main/java/com/example/boot/model/Account.java
@@ -1,5 +1,6 @@
 package com.example.boot.model;
 
+import java.math.BigDecimal;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
@@ -26,7 +27,7 @@ public class Account {
     @Enumerated(EnumType.STRING)
     private Currency currency;
 
-    private double balance;
+    private BigDecimal balance;
 
     @Column(name = "is_active")
     private boolean isActive;
